Plex - Set a timeout of 10 seconds to connect to the server (#596)
Set a timeout of 10 seconds to connect to the server. On connect all known addresses of the server are tried out and afterwards the best one (SSL over non SSL and local over remote) is used. The current timeout sometimes takes longer than the MA timeout of 30 seconds for the setup of a provider