[soundcloud] Improved playlist iteration (#924)
* [soundcloud] improved playlist imports

Similar to https://github.com/music-assistant/server/pull/922, the
playlist import process is incomplete. Through trial and error, I
discovered that soundcloud returns all forms of "track lists" as a
response to the API call -- and that it was sorted by recency (which
explained the behavior I witnessed on `2023.6.0b6`.

This PR fixes it right up. It generalizes soundcloud pagination for
easier future adoption and for the soundcloud playlists.

Manually tested the `asyncsoundcloudpy` API.

* added self to manifest codeowners