Only build mappings when the information is present
When building mappings for later lookup the provider was assuming that
the id and name fields for these mappings would always be present and
set. The API allows for tracks without aritsts or albums and albums
without aritsts so we need to handle these cases. Check if the field has
been set before trying to build the mapping.

Signed-off-by: Eric B Munson <eric@munsonfam.org>