Subsonic: Fix stream scrobbling
get_audio_stream can be called multiple times when a track is played to
allow for MA to analyze the audio. Scrobbling at the end of the stream
was causing multiple reports when placed here. The correct way to do is
is using a callback on the StreamDetails object we create. Now we
scrobble playback when MA tells us that more than half the track has
been played.

We also now make a "Now Playing" scrobble when the StreamDetails object
is built.

Signed-off-by: Eric B Munson <eric@munsonfam.org>