Don't fail entire sync if mbid is corrupt
Have a case where a musicbrainz id on an album is incorrect, and it was blocking my entire library from importing. Luckily MA already catches the error and raises an InvalidDataError on the Artist model setter. This just catches it, so that this artist is imported without the broken data and the other artists are still imported correctly.