Change logger used for provider
This provider was using a mix of self.logger as well as creating it's
own outside the MA infrastructure. The correct thing to do is have the
provider log to self.logger so change all instances to that. This way
all provider logging respects the log level set either globally or for
this provider.

Signed-off-by: Eric B Munson <eric@munsonfam.org>