Fix: Do not enforce cache for every stream
ffmpeg is reading faster than the cache file gets written, causing streams to fail.
Waiting until the file is available, delays all streaming.
Revisit this later. For now only use caching when enforced by provider (apple music) or when we are preloading the next track so when you start a new track, speed is always preferred