Feature: Subsonic: Add bookmark tracking for podcast episodes
The open subsonic spec supports maintaining a play position mid-episode
but we haven't ever wired it up. With the support that has been added to
the provider base class, we now have a simple way to make use of this
functionality. We will create/update a bookmark for a podcast episode
every 30 seconds and when it is stopped. We will delete this bookmark
when an episode has beem fully played.

When an episode is started, we check to see if we have a bookmark and
resume from that point.

Signed-off-by: Eric B Munson <eric@munsonfam.org>