Attempt to fix "This Device" disappearing and reappearing on Firefox
Previously the Builtin Web Player still timed out sometimes when using
Firefox.
This PR increases the previous headroom of 30s to 60s, which should be
enough for Firefox throttling JavaScript in the background.

But I couldn't verify this fix myself since it already worked before on
all devices I have.