Add podcast functionality to Spotify
Has been tested to work locally. 

Only obvious issue is when adding a podcast to the library in MA it doesn't get added to the Spotify library. I am assuming there is some call elsewhere in the MA codebase that isn't getting called for podcasts? All other Spotify media types get added and removed from Spotfy without issue.

There is still some debug code in there as there were problems getting the streams to start but that has settled down so maybe it was a Spotify issue. So let me know if that should be removed.

Otherwise, it is working well. Caching is used to speed up the UI episode lists. Played status is pulled from Spotify.