Fix stop command on some Home Assistant players
Music Assistant attempts to detect the case where a media_stop
command issued to a Home Assistant media player returns an error
because the media player does not support it.

However, the error message that Home Assistant emits has changed
since this check was originally added to Music Assistant.  In this
commit:

https://github.com/home-assistant/core/commit/3aae9b629fc8ca0b8cbf8a2981f0bce20a78e16f

It was changed from:

>  "Entity {entity.entity_id} does not support this service."

To:

> "Entity {entity_id} does not support action {domain}.{service}."

To correct this and support both versions, the substring check is
reduced to the common substring "does not support".