audiobookshelf: Implement more efficient multi-file seeking
The previous seeking implementation relied on downloading all audiobook
files and concatenating them together (and then doing it again every
time the user seeked). With a large book, this very quickly became slow
and a resource hog.

A more efficient way is to use the audio track duration information to
determine which track to start the playback at.