Fix invalid timeout value for TLS connections
aiohttp/connector.py uses the total timeout value as handshake
timeout for TLS connections:

```python
tls_transport = await self._loop.start_tls(
    [...]
    ssl_handshake_timeout=timeout.total,
```

This causes an exception with Deezer on slow network connections:
> ssl_handshake_timeout should be a positive number, got 0

According to aiohttp, a zero value for total is not supposed to be
used, but None means no timeout.

https://github.com/aio-libs/aiohttp/issues/11859#issuecomment-3677980105

Fix it in all occurences throughout the repository.